Add docker-compose file and SHTML support
Switch to using the official Nginx APT repo to get a newer version
of nginx with support for Server Side Includes. This involved some
adjustments to how the site config is deployed (vanilla nginx does
not configure the sites-available/sites-enabled directories). Also
add index.shtml to the list of configured index files.

Also add a docker-compose.yml file as an example of how to run this
with a mapped www directory along with nginx logs, on port 8010.